@@ -9,7 +9,7 @@ return RectorConfig::configure()
         __DIR__ . '/src',
     ])
     // uncomment to reach your current PHP version
-    // ->withPhpSets()
+    ->withPhpSets(php82: true)
     ->withTypeCoverageLevel(74) //49
     ->withDeadCodeLevel(74) // 48
     ->withCodeQualityLevel(74); //74
