@@ -63,7 +63,8 @@ class RolesController extends Controller
      */
     public function actionUpdate(string $name): Response|array|string
     {
-        if ($role = !Role::find($name) instanceof \deadmantfa\yii2\rbac\models\Role) {
+        $role = Role::find($name);
+        if (!$role instanceof Role) {
             throw new NotFoundHttpException('The requested page does not exist.');
         }
 
@@ -90,7 +91,8 @@ class RolesController extends Controller
      */
     public function actionDelete(string $name): Response
     {
-        if ($role = !Role::find($name) instanceof \deadmantfa\yii2\rbac\models\Role) {
+        $role = Role::find($name);
+        if (!$role instanceof Role) {
             throw new NotFoundHttpException('The requested page does not exist.');
         }
 
